% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{get_label}
\alias{get_label}
\title{Title}
\usage{
get_label(
  string,
  language = "en",
  endpoint = "Wikidata",
  labelling_prop = "rdfs:label"
)
}
\arguments{
\item{string}{the string (a part of a triple pattern) to label}

\item{language}{the language in which to return the label (defaults to "en")}

\item{endpoint}{the SPARQL endpoint that is being queried (defaults to "Wikidata")}

\item{labelling_prop}{the name of the labelling property, for instance "skos:prefLabel". Defaults to "rdfs:label"
This function takes a component of a triple pattern as input and returns (if it exists) a corresponding human-readable label.}
}
\value{
the label corresponding to the string
}
\description{
Title
}
\examples{
get_label("wd:Q152088",language="en")
get_label("'David Bowie'")
get_label('"John Doe"')
get_label("?item")
}
