% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query.R
\name{plot_query}
\alias{plot_query}
\title{Plot query as a graph.}
\usage{
plot_query(
  query,
  labelling = FALSE,
  language = "en",
  set_labels = NULL,
  layout = "tree",
  flip = FALSE,
  legend = FALSE,
  show_step = FALSE
)
}
\arguments{
\item{query}{a query as returned by glitter (before call to spq_perform)}

\item{labelling}{whether to label the identifiers in the graph}

\item{language}{the language in which to label. Defaults to English ("en")}

\item{set_labels}{a named vector to replace some elements with a user-defined label. For instance, c("hal:structure"="id de labo")}

\item{layout}{type of layout. Can be e.g. "fr" or "tree". Defaults to "tree"}

\item{flip}{whether to flip the coordinates. Defaults to FALSE}

\item{legend}{whether to draw a legend. Defaults to FALSE}
}
\value{
a plot of the graph depicting query
}
\description{
Plot query as a graph.
}
