% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query_step.R
\name{plot_query_step}
\alias{plot_query_step}
\title{Plot query step as a graph.}
\usage{
plot_query_step(graph, plotstep, layout = "tree", flip = FALSE, legend = FALSE)
}
\arguments{
\item{graph}{a tidygraph object}

\item{plotstep}{the step to plot}

\item{layout}{type of layout. Can be e.g. "fr" or "tree". Defaults to "tree"}

\item{flip}{whether to flip the coordinates. Defaults to FALSE}

\item{legend}{whether to draw a legend. Defaults to FALSE}
}
\value{
a plot of the graph depicting query setp
}
\description{
Plot query step as a graph.
}
\examples{
query=spq_init() \%>\%
  spq_add("?mayor wdt:P31 ?species") \%>\%
  spq_set(species = c('wd:Q144','wd:Q146', 'wd:Q780')) \%>\%
  spq_add("?mayor p:P39 ?node") \%>\%
  spq_add("?node ps:P39 wd:Q30185") \%>\%
  spq_add("?node pq:P642 ?place") 

gq=graph_query(query,labelling=TRUE)
plot_query_step(gq$graph,plotstep=3)
}
