% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_query.R
\name{graph_query}
\alias{graph_query}
\title{A graph depicting the query}
\usage{
graph_query(query, labelling = FALSE, language = "en", set_labels = NULL)
}
\arguments{
\item{query}{a query as returned by glitter (before call to spq_perform)}

\item{labelling}{whether to label the identifiers in the graph}

\item{language}{the language in which to label. Defaults to English ("en")}

\item{set_labels}{a named vector to replace some elements with a user-defined label. For instance, c("hal:structure"="id de labo")}
}
\value{
a tidygraph
}
\description{
A graph depicting the query
}
